Proxy calls only for dev or use playback directly
- Proxy calls in dev environments
- Pass calls directly to playback for internal request
