+++ naturblick/prodsettings.py
@@ -58,3 +58,5 @@ CSRF_COOKIE_SECURE = True
 SESSION_COOKIE_SECURE = True
 CSRF_TRUSTED_ORIGINS = ['https://staging.naturblick.museumfuernaturkunde.berlin',
                         'https://naturblick.museumfuernaturkunde.berlin']
+
+PLAYBACK_URL="http://playback:9000/"
