+++ web/templates/partials/audio.html
@@ -10,9 +10,9 @@
     </div>
 
     <div class="spectrogram_container" id="spectrogram_container">
-        <img src="{{ png|safe }}">
+        <img src="{{ png }}">
         <div class="cover" id="widther" style="width: 0%;">
-            <img src="{{ png|safe }}">
+            <img src="{{ png }}">
         </div>
     </div>
 
@@ -22,7 +22,7 @@
     </div>
 
     <audio id="audio">
-        <source src="{{ url|safe }}">
+        <source src="{{ url }}">
     </audio>
 </div>
